Few changes in mingrid, many in terminal client
Changed border in minegrid to simply be #. Also, now Reveal and
ToggleFlag return whether or not anything changed. Finally, each
MineGrid tracks its state--whether or not the game is still going.

Drastic changes to the terminal client. The game has several
and each input is handled by a function which is looked up in a
table. The display is mostly shared with a few special cases
depending on the state.